Day 16

+++ AdventOfCode/Days/Day12.cs
@@ -2,41 +2,41 @@ namespace AdventOfCode.Days;
 
 public class ConditionRecord
 {
-    private readonly List<long> _springConditions;
+    private readonly List<int> _springConditions;
     private readonly char[] _conditionArray;
     
-    public ConditionRecord(string inputRow, long part = 1)
+    public ConditionRecord(string inputRow, int part = 1)
     {
         var repeat = part == 1 ? 1 : 5;
         var split = inputRow.Split(' ');
-        var numbers = split[1].Split(',').Select(x => Convert.ToInt64(x)).ToList();
+        var numbers = split[1].Split(',').Select(x => Convert.ToInt32(x)).ToList();
         _springConditions = Enumerable.Repeat(numbers, repeat).SelectMany(x => x).ToList();
         _conditionArray = string.Join('?', Enumerable.Repeat(split[0], repeat)).ToCharArray();
     }
 
-    private bool CanPlaceAt(long index, long hashLength)
+    private bool CanPlaceAt(int index, int hashLength)
     {
         if (index + hashLength == _conditionArray.Length)
         {
-            return _conditionArray.Skip((int)index).Take((int)hashLength).All(x => x != '.');
+            return _conditionArray.Skip(index).Take(hashLength).All(x => x != '.');
         }
         if (index + hashLength > _conditionArray.Length)
         {
             return false; 
         }
 
-        return _conditionArray.Skip((int)index).Take((int)hashLength).All(x => x != '.') 
+        return _conditionArray.Skip(index).Take(hashLength).All(x => x != '.') 
                && _conditionArray[index + hashLength] != '#';
     }
     
 
     public long Combinations()
     {
-        var potentials = new List<(long index, long times)> { (0,1) };
+        var potentials = new List<(int index, long times)> { (0,1) };
         
         foreach (var springCondition in _springConditions)
         {
-            var newPotentials = new List<(long index, long times)>();
+            var newPotentials = new List<(int index, long times)>();
             foreach (var (index, times) in potentials)
             {
                 newPotentials.AddRange(FindPotentialSpots(index, springCondition, times));
@@ -48,14 +48,14 @@ public class ConditionRecord
                 .ToList();
         }
 
-        return potentials.Where(x => _conditionArray.Skip((int)x.index).All(ch => ch != '#')).Select(x => x.times).Sum();
+        return potentials.Where(x => _conditionArray.Skip(x.index).All(ch => ch != '#')).Select(x => x.times).Sum();
 
     }
 
-    private IEnumerable<(long index, long time)> FindPotentialSpots(long index, long springCondition, long combos)
+    private IEnumerable<(int index, long time)> FindPotentialSpots(int index, int springCondition, long combos)
     
     {
-        var potentials = new List<long>();
+        var potentials = new List<int>();
         while (true)
         {
             if (index >= _conditionArray.Length || index > 0 && _conditionArray[index - 1] == '#')
